+++ R/seqArchRplus_motif_heatmaps.R
@@ -48,12 +48,13 @@
 #'          package = "seqArchRplus", mustWork = TRUE))
 #'
 #' plot_motif_heatmaps(sname = "sample1", seqs = raw_seqs,
-#'                     flanks = c(50, 100, 200),
+#'                     flanks = c(10, 20, 50),
 #'                     clusts = use_clusts,
 #'                     motifs = c("WW", "SS", "TATAA", "CG", "Y"),
 #'                     dir_path = tempdir(),
 #'                     fheight = 800, fwidth = 1600)
 #'
+#' @author Sarvesh Nikumbh
 plot_motif_heatmaps <- function(sname, seqs, flanks = c(50), clusts,
                                 use_colors = NULL, motifs, dir_path,
                                 fheight = 500, fwidth = 500, funits = "px",